requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
llama-index
sentence-transformers
pymupdf
yfinance
pandas
numpy
sec-edgar-downloader
beautifulsoup4
python-dotenv
pymupdf
nltk
langchain
textstat
langchain-text-splitters

